refactor: consolidate file ingestion pipeline
